push

+++ 006_seg_mrs.py
@@ -63,7 +63,8 @@ def get_mrs_masks(population, afs, workspace_dir):
                         if file.endswith('rda') and 'SUPP' in file:
                             if any(string in file for string in string_list):
                                 print os.path.join(root, file)
-                                shutil.copy(os.path.join(root, file),  os.path.join(vox_dir, voxel))
+                                if os.path.isfile(os.path.join(root, file)):
+                                    shutil.copy(os.path.join(root, file),  os.path.join(vox_dir, voxel))
 
                 # Convert correct RDA
                 matlab_cmd = ['matlab', '-nodesktop', '-nosplash', '-nojvm', '-r "RDA_TO_NIFTI(\'%s\', \'%s\', \'%s\', \'%s\') ; quit;"'
